requirements.txt
django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
Pillow==10.1.0
django-environ==0.11.2
requests